Module select::predicate

Structs

And
Matches if the inner Predicate A and B both match the Node.

Any
Matches any Node.

Attr
Matches Element Node containing attribute N with value V if V is an &str, or any value if V is ().

Class
Matches Element Node containing class T.

Comment
Matches any Comment Node.

Element
Matches any Element Node.

Name
Matches Element Node with name T.

Not
Matches if the Predicate T does not match.

Or
Matches if either inner Predicate A or B matches the Node.

Text
Matches any Text Node.

Traits

Predicate
A trait implemented by all Node matchers.